feat: users/stocks Entity 업데이트 (#11)
  - User: nickname, loginType, isActive 등 컬럼 추가
  - Stock: market 추가, UNIQUE → (code, market) 복합                                                                                                                             - StockPrice: 가격 타입 Long → BigDecimal
  - ErrorCode: DUPLICATE_NICKNAME, STOCK_NOT_FOUND 추가

+++ pong-to-rich/backend/src/main/java/com/pongtorich/pong_to_rich/dto/stock/StockPriceResponse.java
@@ -2,14 +2,15 @@ package com.pongtorich.pong_to_rich.dto.stock;
 
 import com.pongtorich.pong_to_rich.domain.stock.StockPrice;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record StockPriceResponse(
         LocalDate tradeDate,
-        Long openPrice,
-        Long highPrice,
-        Long lowPrice,
-        Long closePrice,
+        BigDecimal openPrice,
+        BigDecimal highPrice,
+        BigDecimal lowPrice,
+        BigDecimal closePrice,
         Long volume
 ) {
     public static StockPriceResponse from(StockPrice stockPrice) {
